cambio int to float en posiciones de los barcos.

+++ backend/src/logica/entidades/Patrulla.java
@@ -20,7 +20,7 @@ public class Patrulla extends Vehiculo {
 	}
 
 	// Constructor para usar cuando se recupera de la BD
-	public Patrulla(int id, int x, int y, String tipo, String tipoPatrulla, float initialRotation, float combustible) {
+	public Patrulla(int id, float x, float y, String tipo, String tipoPatrulla, float initialRotation, float combustible) {
 		Properties p = new Properties();
 		InputStream input = null;
 		
@@ -104,7 +104,7 @@ public class Patrulla extends Vehiculo {
 		}
 	}
 
-	public Patrulla(int id, int x, int y, float velocidad, float velocidadAngular, String spriteVivo, String spritesLaterales, float combustible, Armas armas, float initialRotation, String tipo, String tipoPatrulla, int gastoCombustible) {
+	public Patrulla(int id, float x, float y, float velocidad, float velocidadAngular, String spriteVivo, String spritesLaterales, float combustible, Armas armas, float initialRotation, String tipo, String tipoPatrulla, int gastoCombustible) {
 		super(id, x, y, tipo, velocidad, velocidadAngular, spriteVivo, spritesLaterales, initialRotation);
 		this.combustible = combustible;
 		this.setArmas(armas);
