+++ backend/src/logica/entidades/Vehiculo.java
@@ -6,8 +6,8 @@ import com.google.gson.JsonElement;
 public abstract class Vehiculo {
 
 	private int id;
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 	private String tipo;
 	private float velocidad;
 	private float velocidadAngular;
@@ -19,7 +19,7 @@ public abstract class Vehiculo {
 		
 	}
 	
-	public Vehiculo(int id, int x, int y, String tipo, float velocidad, float velocidadAngular, String spriteVivo, String spritesLaterales, float initialRotation) {
+	public Vehiculo(int id, float x, float y, String tipo, float velocidad, float velocidadAngular, String spriteVivo, String spritesLaterales, float initialRotation) {
 		this.id = id;
 		this.x = x;
 		this.y = y;
@@ -43,7 +43,7 @@ public abstract class Vehiculo {
 		return x;
 	}
 	
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 	
@@ -51,7 +51,7 @@ public abstract class Vehiculo {
 		return y;
 	}
 	
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
